+++ src/assets.ts
@@ -7,6 +7,7 @@ export enum Asset {
   BTC = "BTC",
   ETH = "ETH",
   APT = "APT",
+  ARB = "ARB",
   UNDEFINED = "UNDEFINED",
 }
 
@@ -45,6 +46,7 @@ export function assetToName(asset: Asset): string | null {
   if (asset == Asset.BTC) return "BTC";
   if (asset == Asset.ETH) return "ETH";
   if (asset == Asset.APT) return "APT";
+  if (asset == Asset.ARB) return "ARB";
   if (asset == Asset.UNDEFINED) return "UNDEFINED";
   if (asset == null) return null; // Some things, like clock callbacks, are for all assets and return asset=null
   throw Error("Invalid asset");
@@ -55,6 +57,7 @@ export function nameToAsset(name: string): Asset {
   if (name == "BTC") return Asset.BTC;
   if (name == "ETH") return Asset.ETH;
   if (name == "APT") return Asset.APT;
+  if (name == "ARB") return Asset.ARB;
   if (name == "UNDEFINED") return Asset.UNDEFINED;
   throw Error("Invalid asset");
 }
@@ -68,6 +71,7 @@ export function toProgramAsset(asset: Asset): any {
   if (asset == Asset.BTC) return { btc: {} };
   if (asset == Asset.ETH) return { eth: {} };
   if (asset == Asset.APT) return { apt: {} };
+  if (asset == Asset.ARB) return { arb: {} };
   throw Error("Invalid asset");
 }
 
@@ -83,6 +87,9 @@ export function fromProgramAsset(asset: any): Asset {
   }
   if (objectEquals(asset, { apt: {} })) {
     return Asset.APT;
+  }
+  if (objectEquals(asset, { arb: {} })) {
+    return Asset.ARB;
   }
   throw Error("Invalid asset");
 }
@@ -101,6 +108,9 @@ export function assetToIndex(asset: Asset): number {
     case Asset.APT: {
       return 3;
     }
+    case Asset.ARB: {
+      return 4;
+    }
   }
   throw new Error("Invalid asset");
 }
@@ -118,6 +128,9 @@ export function indexToAsset(index: number): Asset {
     }
     case 3: {
       return Asset.APT;
+    }
+    case 4: {
+      return Asset.ARB;
     }
   }
   throw new Error("Invalid index");