+++ src/types/zeta.ts
@@ -592,6 +592,11 @@ export type Zeta = {
           "name": "admin",
           "isMut": true,
           "isSigner": true
+        },
+        {
+          "name": "secondaryAdmin",
+          "isMut": false,
+          "isSigner": false
         }
       ],
       "args": [
@@ -687,6 +692,27 @@ export type Zeta = {
     },
     {
       "name": "updateAdmin",
+      "accounts": [
+        {
+          "name": "state",
+          "isMut": true,
+          "isSigner": false
+        },
+        {
+          "name": "admin",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "newAdmin",
+          "isMut": true,
+          "isSigner": true
+        }
+      ],
+      "args": []
+    },
+    {
+      "name": "updateSecondaryAdmin",
       "accounts": [
         {
           "name": "state",
@@ -4790,11 +4816,19 @@ export type Zeta = {
             "type": "u8"
           },
           {
+            "name": "maxPerpDeltaAge",
+            "type": "u16"
+          },
+          {
+            "name": "secondaryAdmin",
+            "type": "publicKey"
+          },
+          {
             "name": "padding",
             "type": {
               "array": [
                 "u8",
-                107
+                73
               ]
             }
           }
@@ -5991,6 +6025,10 @@ export type Zeta = {
           {
             "name": "marginConcessionPercentage",
             "type": "u8"
+          },
+          {
+            "name": "maxPerpDeltaAgeSeconds",
+            "type": "u16"
           }
         ]
       }
@@ -6087,6 +6125,10 @@ export type Zeta = {
           {
             "name": "marginConcessionPercentage",
             "type": "u8"
+          },
+          {
+            "name": "maxPerpDeltaAgeSeconds",
+            "type": "u16"
           }
         ]
       }
@@ -6542,6 +6584,9 @@ export type Zeta = {
           },
           {
             "name": "APT"
+          },
+          {
+            "name": "ARB"
           },
           {
             "name": "UNDEFINED"
@@ -8225,6 +8270,11 @@ export const IDL: Zeta = {
           "name": "admin",
           "isMut": true,
           "isSigner": true
+        },
+        {
+          "name": "secondaryAdmin",
+          "isMut": false,
+          "isSigner": false
         }
       ],
       "args": [
@@ -8320,6 +8370,27 @@ export const IDL: Zeta = {
     },
     {
       "name": "updateAdmin",
+      "accounts": [
+        {
+          "name": "state",
+          "isMut": true,
+          "isSigner": false
+        },
+        {
+          "name": "admin",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "newAdmin",
+          "isMut": true,
+          "isSigner": true
+        }
+      ],
+      "args": []
+    },
+    {
+      "name": "updateSecondaryAdmin",
       "accounts": [
         {
           "name": "state",
@@ -12423,11 +12494,19 @@ export const IDL: Zeta = {
             "type": "u8"
           },
           {
+            "name": "maxPerpDeltaAge",
+            "type": "u16"
+          },
+          {
+            "name": "secondaryAdmin",
+            "type": "publicKey"
+          },
+          {
             "name": "padding",
             "type": {
               "array": [
                 "u8",
-                107
+                73
               ]
             }
           }
@@ -13624,6 +13703,10 @@ export const IDL: Zeta = {
           {
             "name": "marginConcessionPercentage",
             "type": "u8"
+          },
+          {
+            "name": "maxPerpDeltaAgeSeconds",
+            "type": "u16"
           }
         ]
       }
@@ -13720,6 +13803,10 @@ export const IDL: Zeta = {
           {
             "name": "marginConcessionPercentage",
             "type": "u8"
+          },
+          {
+            "name": "maxPerpDeltaAgeSeconds",
+            "type": "u16"
           }
         ]
       }
@@ -14175,6 +14262,9 @@ export const IDL: Zeta = {
           },
           {
             "name": "APT"
+          },
+          {
+            "name": "ARB"
           },
           {
             "name": "UNDEFINED"